+++ app/src/main/java/com/example/soft1c/model/SizeAcceptance.kt
@@ -5,7 +5,7 @@ data class SizeAcceptance(
     val sum: Int = 0,
     val allWeight: Int = 0,
     val priceM3: Int = 0,
-    val priceWeight: Int = 0,
+    val priceWeight: Double = 0.0,
     var dataArray: List<SizeData>,
 ) {
 
@@ -14,7 +14,7 @@ data class SizeAcceptance(
         var length: Int = 0,
         var width: Int = 0,
         var height: Int = 0,
-        var weight: Int = 0,
+        var weight: Double = 0.0,
     )
 
 }
